refactor: establish new project structure
